Moved storage management to separate subsystem for clarity
